Add better handling for logins, edit using directives.
Adds the new encryption service via the package. Also checks for the same booleans as the register page, so rather than calling the api with a username or password that can't even be registered, it doesn't call the api at all and instead just tells the user that the username or password is incorrect, since it'd be impossible for someone to have that username and or password.