Fix bug: Getting user projects errored out if the user had no projects
Rather than making a list if the user has projects and then returning that list, make the list after the request is sent. Add the body result to the list if the user had no projects and add the projects to the list if the user does have projects, and then return the list. Initially, the list was empty if there were no projects found but the list was being returned anyway which caused the error.